+++ packages/cf-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 1.0.32
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/inquirer-common 1.0.28 → 1.0.29
+
 ## 1.0.31
 
 ### Patch Changes